Add cli-bin mode
