+++ src/index.ts
@@ -1,5 +1,4 @@
 
-import * as fs from 'fs'
 import csvParse from 'csv-parse/lib/sync'
 
 // Import data types
@@ -7,7 +6,6 @@ import { KDBData, KDBCourse } from '../types'
 
 // Import utils
 import * as arrayUtil from './util/array'
-const Iconv = require('iconv').Iconv
 
 // Export data types
 export { KDBData, KDBCourse }
@@ -145,20 +143,4 @@ export default function parse (csvData: string): KDBData {
   })
 
   return output
-}
-
-if (require.main === module) {
-  const iconv = new Iconv('Shift_JIS', 'UTF-8//TRANSLIT//IGNORE')
-
-  const filename = process.argv[ 2 ]
-  if (!filename) {
-    console.error('ERROR: Filename is required')
-    process.exit(1)
-  }
-
-  const csvData = iconv.convert(fs.readFileSync(filename)).toString()
-
-  const result = parse(csvData)
-
-  process.stdout.write(JSON.stringify(result))
 }
